TST/REF: method specific test files
