+++ pandas/tests/series/test_api.py
@@ -163,10 +163,7 @@ class TestSeriesMisc:
             assert val == string_series[i]
 
     def test_keys(self, datetime_series):
-        # HACK: By doing this in two stages, we avoid 2to3 wrapping the call
-        # to .keys() in a list()
-        getkeys = datetime_series.keys
-        assert getkeys() is datetime_series.index
+        assert datetime_series.keys() is datetime_series.index
 
     def test_values(self, datetime_series):
         tm.assert_almost_equal(
@@ -212,10 +209,6 @@ class TestSeriesMisc:
         # https://github.com/pandas-dev/pandas/issues/18147
         # no exception and no empty docstring
         assert pydoc.getdoc(Series.index)
-
-    def test_numpy_unique(self, datetime_series):
-        # it works!
-        np.unique(datetime_series)
 
     def test_item(self):
         s = Series([1])
